Update README.md
This project represents my fourth venture into web development, focusing on HTML and CSS. Throughout this practice project, I delved into the fundamentals of creating a responsive and visually appealing web page. One of the highlights of this endeavor is the integration of the HTML5 '<details>' and '<summary>' elements to enhance user interaction and content organization.